Refactor log_config utility
