+++ examples/bert/main.py
@@ -3,11 +3,10 @@
 
 import os
 import sys
-from typing import Dict, Optional, cast
+from typing import Optional, cast
 
-import wandb
 from composer import Trainer
-from composer.utils import dist, reproducibility
+from composer.utils import reproducibility
 from omegaconf import DictConfig
 from omegaconf import OmegaConf as om
 
@@ -16,7 +15,7 @@ from examples.bert.src.mosaic_bert import create_mosaic_bert_mlm
 from examples.common.builders import (build_algorithm, build_callback,
                                       build_dataloader, build_logger,
                                       build_optimizer, build_scheduler)
-from examples.common.config_utils import update_batch_size_info
+from examples.common.config_utils import log_config, update_batch_size_info
 
 
 def build_model(cfg: DictConfig):
@@ -123,16 +122,7 @@ def main(cfg: DictConfig,
     )
 
     print('Logging config...')
-    config_dict = om.to_container(cfg, resolve=True)
-    assert isinstance(config_dict, (Dict,))  # type checking
-    config_dict.update({
-        'n_gpus': dist.get_world_size(),
-        'n_params': n_params,
-        'device_train_batch_size': cfg.device_train_batch_size,
-        'device_eval_batch_size': cfg.device_eval_batch_size,
-    })
-    if wandb.run is not None:
-        wandb.config.update(config_dict)
+    log_config(cfg)
 
     if do_train:
         print('Starting training...')
